+++ task1.c
@@ -5,14 +5,14 @@
 #include "process.h"
 #include "queue.h"
 
-void round_robin(process_t **processes, queue_t *queue, int num_process, int quantum, int *makespan)
+void round_robin(process_t **processes, queue_t *queue, int num_process, int quantum, unsigned long int *makespan)
 {
     /*
         scheduler: allocates processes in queue, CPU time for one quantum if they have a memory allocation,
             if not then they are sent to the back of the queue. implements roundrobin algorithm
     */
 
-    int simulation_time = 0;
+    unsigned long int simulation_time = 0;
     int submitted_process = 0;
     int remaining_process = num_process;
     int ready_process_remaining = 0;
@@ -41,7 +41,7 @@ void round_robin(process_t **processes, queue_t *queue, int num_process, int qua
                 remaining_process--;
                 ready_process_remaining = get_queue_length(queue);
 
-                printf("%d,%s,process-name=%s,proc-remaining=%d\n", simulation_time, get_status_string(current_process), current_process->name, ready_process_remaining);
+                printf("%lu,%s,process-name=%s,proc-remaining=%d\n", simulation_time, get_status_string(current_process), current_process->name, ready_process_remaining);
 
                 current_process->completion_time = simulation_time; //  time of completion for the process
                 current_process->turnaround_time = simulation_time - current_process->arrival_time;
@@ -77,7 +77,7 @@ void round_robin(process_t **processes, queue_t *queue, int num_process, int qua
             current_process = dequeue(queue);
             change_status(current_process, RUNNING);
 
-            printf("%d,%s,process-name=%s,remaining-time=%lu\n", simulation_time, get_status_string(current_process), current_process->name, current_process->remaining_time);
+            printf("%lu,%s,process-name=%s,remaining-time=%lu\n", simulation_time, get_status_string(current_process), current_process->name, current_process->remaining_time);
         }
 
         if (current_process != NULL && current_process->remaining_time > 0)
