+++ task1.h
@@ -7,6 +7,6 @@
 #define TRUE 1
 #define FALSE 0
 
-void round_robin(process_t **processes, queue_t *queue, int num_process, int quantum, int *makespan);
+void round_robin(process_t **processes, queue_t *queue, int num_process, int quantum, unsigned long int *makespan);
 
 #endif
