+++ task3.c
@@ -24,14 +24,14 @@ void print_process3(process_t *process)
            process->name, process->arrival_time, process->service_time, process->memory_KB, process->page_table->amount);
 }
 
-void paged_scheduler(process_t **processes, queue_t *queue, int num_processes, int quantum, int *makespan, allocation_t *allocation)
+void paged_scheduler(process_t **processes, queue_t *queue, int num_processes, int quantum, unsigned long int *makespan, allocation_t *allocation)
 {
     /*
         scheduler: allocates processes in queue CPU time for one quantum if they have a memory allocation,
             if not then they are sent to the back of the queue. uses roundrobin algorithm
     */
 
-    int simulation_time = 0;
+    unsigned long int simulation_time = 0;
     int submitted_process = 0;
     int remaining_process = num_processes;
     int ready_process_remaining = 0;
@@ -67,7 +67,7 @@ void paged_scheduler(process_t **processes, queue_t *queue, int num_processes, i
 
                 deallocate_allocation(allocation, current_process->page_table, current_process->id, simulation_time);
                 print_eviction(allocation, simulation_time);
-                printf("%d,%s,process-name=%s,proc-remaining=%d\n", simulation_time, get_status_string(current_process), current_process->name, ready_process_remaining);
+                printf("%lu,%s,process-name=%s,proc-remaining=%d\n", simulation_time, get_status_string(current_process), current_process->name, ready_process_remaining);
                 current_process = NULL;
 
                 if (remaining_process == 0 && is_empty(queue))
@@ -98,7 +98,7 @@ void paged_scheduler(process_t **processes, queue_t *queue, int num_processes, i
                 change_status(current_process, RUNNING);
 
                 percentage = ceil((100 * (float)((allocation->size) - (allocation->vacancies)) / allocation->size));
-                printf("%d,%s,process-name=%s,remaining-time=%lu,mem-usage=%.0f%%,",
+                printf("%lu,%s,process-name=%s,remaining-time=%lu,mem-usage=%.0f%%,",
                        simulation_time, get_status_string(current_process), current_process->name, current_process->remaining_time,
                        percentage);
                 print_table(current_process->page_table);
@@ -122,7 +122,7 @@ void paged_scheduler(process_t **processes, queue_t *queue, int num_processes, i
     *makespan = simulation_time;
 }
 
-void evict_and_allocate(allocation_t *allocation, process_t **processes, int num_processes, process_t *process, int time)
+void evict_and_allocate(allocation_t *allocation, process_t **processes, int num_processes, process_t *process, unsigned long int time)
 {
     // step 1: evict process from pages and allocate pages to new process
     process_t *evicted = NULL;
@@ -141,7 +141,7 @@ void evict_and_allocate(allocation_t *allocation, process_t **processes, int num
     allocate_pages(allocation, process->page_table, process->id);
 }
 
-process_t *least_recently_executed(process_t **processes, int num_processes, int time)
+process_t *least_recently_executed(process_t **processes, int num_processes, unsigned long int time)
 {
     /*
         linear search to find the least recently executed
@@ -161,7 +161,7 @@ process_t *least_recently_executed(process_t **processes, int num_processes, int
     return temp;
 }
 
-process_t *get_next_paged_process(queue_t *queue, allocation_t *allocation, process_t **processes, int num_processes, int time)
+process_t *get_next_paged_process(queue_t *queue, allocation_t *allocation, process_t **processes, int num_processes, unsigned long int time)
 {
     /*
         get next process that has been allocated memory, if no memory allocation and
